Población a nivel comunal de Chile (2024)

Cifras de población según el Censo 2024

Usage

censo_poblacion_2024

Format

Un data frame con 346 filas y 8 columnas:

codigo_region

Código único territorial de las regiones de Chile (número entre 1 y 16)

nombre_region

Nombre de las regiones de Chile

codigo_provincia

Código único territorial de las provincias de Chile

nombre_provincia

Nombre de las provincias de Chile

codigo_comuna

Código único territorial de las comunas de Chile

nombre_comuna

Nombre de las comunas de Chile

año

Año al que corresponde la población

poblacion

Cantidad de habitantes para la comuna y el año correspondientes

Source

https://www.ine.gob.cl/estadisticas-por-tema/demografia-y-poblacion/censo-de-poblacion-y-vivienda

Details

Los datos fueron obtenidos desde el sitio web del INE.